% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-data.R
\docType{data}
\name{classdata.std}
\alias{classdata.std}
\title{Classification Data - Breast Cancer Wisconsin (Diagnostic)}
\format{
A data frame with 569 rows and 31 variables.
}
\source{
http://archive.ics.uci.edu/ml/index.php
}
\usage{
data(classdata.std)
}
\description{
Pre-processed and standardized Breast Cancer Wisconsin (Diagnostic)
is a real-life data set from University of California at Irvine (UCI)
machine learning data repository UCI: Machine Learning Repository
(available at http://archive.ics.uci.edu/ml/).
}
\examples{
\dontrun{
data(classdata.std)
cmaRs(y ~ ., classification = TRUE, data = classdata.std)
}
}
\references{
Dua, D. and Graff, C. (2017). UCI machine learning repository.
}
\keyword{datasets}
