% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cmaRs.R
\name{summary.cmaRs}
\alias{summary.cmaRs}
\title{A summary function designed for CMARS}
\usage{
\method{summary}{cmaRs}(object, ...)
}
\arguments{
\item{object}{A cmaRs object which is constructed by cmaRs.}

\item{...}{Additional arguments affecting the summary result.}
}
\value{
An S3 model of class "summary.cmaRs"
}
\description{
This function allows you to print the output of CMARS model.
}
\examples{
\dontrun{
# Without \code{MOSEK}, the example code is not executable.
# For installation of Mosek, plese see the documentation of 'Rmosek'.
data("trees", package = "datasets")
model.prediction <- cmaRs(Volume ~ ., degree = 5, nk = 20, data = trees)
summary.cmaRs(model.prediction)
}
}
