% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableb6-data.R
\docType{data}
\name{table.b6}
\alias{table.b6}
\title{Data Set for Table B6}
\format{
A data frame with 28 rows and 5 variables.
}
\source{
Montgomery, D.C., Peck, E.A., and Vining, C.G. (2001)
Introduction to Linear Regression Analysis.
3rd Edition, John Wiley and Sons.
}
\usage{
data(table.b6)
}
\description{
This data set is taken from MPV package (version 1.58).
Please check library(MPV) for more detail.
}
\examples{
data(table.b6)
}
\references{
(1972) Kinetics of Chlorination of Niobium oxychloride
by Phosgene in a Tube-Flow Reactor.
Industrial and Engineering Chemistry, Process Design Development, 11(2).
}
\keyword{datasets}
